Split monitoring router and hub into separate files
This is ongoing work to split up monitoring.py into more
topical pieces - see also PRs #2468, #2439
